factored out some code from receptor baseclass into a new WarpBC baseclass
